% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog.R
\name{mylog}
\alias{mylog}
\title{My log function}
\usage{
mylog(x)
}
\arguments{
\item{x}{a numerical vector of length n}
}
\value{
a numerical vector of log(x) of length n
}
\description{
My log function
}
\examples{
mylog(c(10, exp(2), 1))
}
